Create  #Allocate_minimum_pages_of_book.cpp
If there are n different books having a different number of pages and m students. 
The books are arranged in ascending order of a number of pages. Every student is assigned to read some consecutive books. 
The task is to assign books in such a way that the maximum number of pages assigned to a student is minimum. 

~ One book can be assigned to only one student. Two stdents can't read from same book.
~ Every student will definitely get at least One Book. So we can't give all the books to any one student.
~ Books will be distributed in contineous manner to any student. 
~ EX:- Books [A,B,C,D] and 2 students then :-     Student 1 == Student 2
                                                     A           B,C,D
                                                    A,B           C,D 
                                                   A,B,C           D